feat(vortex-geo): prune ST_Intersects filters via the AABB zone statistic
- Add GeoIntersectsPrune: skips chunks whose AABB is strictly separated from
  the filter constant's bounding box (touching boxes scan, since touching
  geometries intersect); NOT-intersects and unprovable shapes fall through
  to the scan
- Restructure prune.rs into a directory module ready for future spatial
  predicates: mod.rs holds the shared proof vocabulary (geometry_and_constant,
  query_aabb, aabb_stat, min_dist_sq/max_dist_sq), each predicate is a leaf
  module owning its own test cases (distance.rs, intersects.rs), and shared
  zone-map fixtures live in test_harness.rs
- Fold in the #8646 review follow-ups that missed its merge: the aabb_stat
  ctor (part of the restructure) and tests pinning the radius-literal
  tolerance set - an integer radius casts and prunes, null and
  extension-typed radii decline to a scan; the harness takes any literal
  scalar
